Other Projects/Plane Game/plane-simulator/requirements.txt
pygame
PyOpenGL
numpy
scipy
Pillow